requirements.txt
flet==0.9.0
pyzbar==0.1.8
opencv-python==4.8.0.74
pyjnius==1.4.2